Relocate to stores/alert.tsx
